Fix all occurances

+++ app/src/main/java/com/example/myapplication/recipes/data/datasource/RecipeDao.kt
@@ -14,7 +14,7 @@ interface RecipeDao {
     fun getRecipes(): Flow<List<Recipe>>
 
     @Query("SELECT * FROM recipe WHERE id =:id")
-    suspend fun getRecipeById(id: Int): Recipe?
+    suspend fun getRecipeById(id: Long): Recipe?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertRecipe(recipe: Recipe): Long
