+++ app/src/main/java/com/example/myapplication/recipes/domain/repository/RecipeRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 interface RecipeRepository {
     fun getRecipes(): Flow<List<Recipe>>
 
-    suspend fun getRecipeById(id: Int): Recipe?
+    suspend fun getRecipeById(id: Long): Recipe?
 
     suspend fun insertRecipe(recipe: Recipe): Long
